backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
python-docx
PyMuPDF
